+++ mrb_doc/core/timing.rb
@@ -1,11 +1,3 @@
-# Returns the current framerate as an integer.
-# @return [Integer]
-def get_fps
-  # mrb_get_fps
-  # src/mruby_integration/core/timing.cpp
-  60
-end
-
 # Returns the amount of time the last frame took in seconds.
 # @return [Float]
 def get_frame_time