+++ src/mruby_integration/core/timing.cpp
@@ -2,12 +2,6 @@
 #include "raylib.h"
 
 auto
-mrb_get_fps(mrb_state* mrb, mrb_value) -> mrb_value
-{
-  return mrb_int_value(mrb, GetFPS());
-}
-
-auto
 mrb_get_frame_time(mrb_state* mrb, mrb_value) -> mrb_value
 {
   return mrb_float_value(mrb, GetFrameTime());
@@ -21,8 +15,6 @@ mrb_get_time(mrb_state* mrb, mrb_value) -> mrb_value
 void
 append_core_timing(mrb_state* mrb)
 {
-  mrb_define_method(
-    mrb, mrb->kernel_module, "get_fps", mrb_get_fps, MRB_ARGS_NONE());
   mrb_define_method(mrb,
                     mrb->kernel_module,
                     "get_frame_time",